# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [household]
      summary: Get the current user, their household and members
      description: >-
        Returns the authenticated user. On first call this JIT-provisions the
        user, a default household, and a membership.
      responses:
        "200":
          description: Current user context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Authenticated user no longer has valid access to the active household
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Active household identity or membership authority is internally inconsistent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateMe
      tags: [household]
      summary: Update the current user's profile / display name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeUpdate"
      responses:
        "200":
          description: Updated user context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400":
          description: Invalid profile update payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Authenticated user no longer has valid access to the active household
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Active household identity or membership authority changed during the update or exact response snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /household/invitations:
    get:
      operationId: listHouseholdInvitations
      tags: [household]
      summary: List household invitations
      description: >-
        Owner/admin-only review of provider-durable household invitation
        lifecycle rows scoped to the authenticated active household. Production
        Supabase migration, RLS, retention, export/deletion policy, and
        notification delivery remain launch approval gates.
      parameters:
        - $ref: "#/components/parameters/ExpectedHouseholdId"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: lifecycleState
          in: query
          schema:
            $ref: "#/components/schemas/HouseholdInvitationLifecycleState"
      responses:
        "200":
          description: Durable household invitation review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseholdInvitationListResponse"
        "400":
          description: Invalid invitation review query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Only owners can review invitations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "412":
          $ref: "#/components/responses/ExpectedHouseholdMismatch"
        "428":
          $ref: "#/components/responses/ExpectedHouseholdRequired"
    post:
      operationId: createHouseholdInvitation
      tags: [household]
      summary: Create a household invitation
      description: >-
        Owner/admin-only creation of provider-durable invitation rows. New
        invitations may start approved or pending owner approval, and the join
        flow only creates caregiver memberships after the row is approved and
        unexpired.
      parameters:
        - $ref: "#/components/parameters/ExpectedHouseholdId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HouseholdInvitationCreateInput"
      responses:
        "201":
          description: Household invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseholdInvitationMutationResponse"
        "400":
          description: Invalid invitation payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Only owners can create invitations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Household authority or invitation state changed before the atomic create completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "412":
          $ref: "#/components/responses/ExpectedHouseholdMismatch"
        "428":
          $ref: "#/components/responses/ExpectedHouseholdRequired"
...
  /household/join:
    post:
      operationId: joinHousehold
      tags: [household]
      summary: Join an existing household by invite code
      description: >-
        Accepts only a provider-durable invitation code, creates or reuses the
        caregiver membership, and returns the persisted HouseholdAuditEvent for
        invitation acceptance. The expected-household capability identifies the
        exact source household and is compared again inside the join transaction.
      parameters:
        - $ref: "#/components/parameters/ExpectedHouseholdId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinHouseholdInput"
      responses:
        "200":
          description: Joined household context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseholdJoinResponse"
        "400":
          description: Invalid invite payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Invitation is not approved for this authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Invite code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Household membership or active authority changed during join
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "412":
          $ref: "#/components/responses/ExpectedHouseholdMismatch"
        "428":
          $ref: "#/components/responses/ExpectedHouseholdRequired"
        "503":
          description: Provider-backed household join storage is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /care-state:
    get:
      operationId: getCareState
      tags: [care-state]
      summary: Get the household's synced care document
      parameters:
        - $ref: "#/components/parameters/ExpectedHouseholdId"
      responses:
        "200":
          description: Care state document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareStateEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Authenticated user has no valid role in the expected household
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Care state document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Household authority became internally inconsistent before the read completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "412":
          $ref: "#/components/responses/ExpectedHouseholdMismatch"
        "428":
          $ref: "#/components/responses/ExpectedHouseholdRequired"
    put:
      operationId: putCareState
      tags: [care-state]
      summary: Replace the household's synced care document
      parameters:
        - $ref: "#/components/parameters/ExpectedHouseholdId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareStateInput"
      responses:
        "200":
          description: Updated care state document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareStateEnvelope"
        "400":
          description: Invalid care state payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Authenticated household role is not allowed to update shared Care state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Care state document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Version conflict or household authority integrity failure; refetch before retrying
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/CareStateEnvelope"
                  - $ref: "#/components/schemas/ApiError"
        "412":
          $ref: "#/components/responses/ExpectedHouseholdMismatch"
        "428":
          $ref: "#/components/responses/ExpectedHouseholdRequired"
...
components:
  parameters:
    ExpectedHouseholdId:
      name: X-WoofWatcher-Expected-Household-Id
      in: header
      required: true
      description: Exact opaque active household identity previously resolved by the authenticated client. Values are compared byte-for-byte and are never trimmed or case-folded.
      schema:
        type: string
        minLength: 1
        pattern: '.*\S.*'
  responses:
    ExpectedHouseholdMismatch:
      description: The authenticated user's active household no longer matches the expected household.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    ExpectedHouseholdRequired:
      description: An exact expected-household capability is required before accessing household-scoped data.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
      required:
        - id
    Household:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        inviteCode:
          type: string
          description: Deprecated compatibility field. Always empty; durable HouseholdInvitation rows are the only production join credentials.
          deprecated: true
          maxLength: 0
      required:
        - id
        - name
        - inviteCode
    HouseholdMemberRole:
      type: string
      description: Canonical household membership role shared by household, invitation, and retained-membership contracts.
      enum: [owner, adult, teen, kid, sitter, trainer, walker, vet viewer]
...
    Member:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        role:
          $ref: "#/components/schemas/HouseholdMemberRole"
        displayName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        isSelf:
          type: boolean
        accessPassExpiresAt:
          type: ["string", "null"]
          description: Active Access Pass expiry for a temporary helper. At this instant all household access ends until an owner grants a new pass.
        accessPassExpired:
          type: boolean
          description: True when a temporary helper's Access Pass has lapsed and writes should be blocked until renewed.
      required:
        - id
        - userId
        - role
        - isSelf
    Me:
      type: object
      properties:
        authorityObservedAt:
          type: string
          format: date-time
          description: Exact transactional database-clock instant used to evaluate Access Pass expiry for this authority snapshot. Clients must pair it with a monotonic request interval and must not substitute device wall-clock time.
        user:
          $ref: "#/components/schemas/User"
        household:
          $ref: "#/components/schemas/Household"
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
      required:
        - authorityObservedAt
        - user
        - household
        - members
    HouseholdAuditEvent:
      type: object
      properties:
        id:
          type: string
        action:
          type: string
          enum:
            - invitation-created
            - invitation-accepted
            - invitation-revoked
            - member-role-updated
            - member-revoked
            - access-pass-activated
            - access-pass-revoked
        lifecycleState:
          type: string
          enum:
            - invite-created
            - invite-accepted
            - invite-revoked
            - member-updated
            - member-revoked
            - access-pass-active
            - access-pass-revoked
            - access-pass-expired
        actorUserId:
          type: string
        householdId:
          type: string
        targetMemberId:
          type: ["string", "null"]
        targetUserId:
          type: ["string", "null"]
        targetRole:
          type: ["string", "null"]
        nextRole:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        storage:
          type: string
          enum: [provider-durable]
        boundary:
          type: string
      required:
        - id
        - action
        - lifecycleState
        - actorUserId
        - householdId
        - createdAt
        - storage
        - boundary
    HouseholdJoinResponse:
      allOf:
        - $ref: "#/components/schemas/Me"
        - type: object
          properties:
            auditEvent:
              $ref: "#/components/schemas/HouseholdAuditEvent"
          required:
            - auditEvent
...
    HouseholdInvitationLifecycleState:
      type: string
      enum:
        - pending-approval
        - approved
        - accepted
        - revoked
        - expired
        - rejected
    HouseholdInvitation:
      type: object
      properties:
        id:
          type: string
        householdId:
          type: string
        inviteCode:
          type: string
        invitedEmail:
          type: ["string", "null"]
        invitedUserId:
          type: ["string", "null"]
        role:
          $ref: "#/components/schemas/HouseholdMemberRole"
        lifecycleState:
          $ref: "#/components/schemas/HouseholdInvitationLifecycleState"
        runtimeLifecycleState:
          $ref: "#/components/schemas/HouseholdInvitationLifecycleState"
        expired:
          type: boolean
        createdByUserId:
          type: string
        approvedByUserId:
          type: ["string", "null"]
        acceptedByUserId:
          type: ["string", "null"]
        revokedByUserId:
          type: ["string", "null"]
        rejectedByUserId:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        acceptedAt:
          type: ["string", "null"]
          format: date-time
        revokedAt:
          type: ["string", "null"]
          format: date-time
        rejectedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
        storage:
          type: string
          enum: [provider-durable]
        boundary:
          type: string
      required:
        - id
        - householdId
        - inviteCode
        - role
        - lifecycleState
        - runtimeLifecycleState
        - expired
        - createdByUserId
        - createdAt
        - storage
        - boundary
    HouseholdInvitationListFilters:
      type: object
      properties:
        lifecycleState:
          $ref: "#/components/schemas/HouseholdInvitationLifecycleState"
    HouseholdInvitationListResponse:
      type: object
      properties:
        invitations:
          type: array
          items:
            $ref: "#/components/schemas/HouseholdInvitation"
        limit:
          type: integer
        filters:
          $ref: "#/components/schemas/HouseholdInvitationListFilters"
        boundary:
          type: string
      required:
        - invitations
        - limit
        - filters
        - boundary
    HouseholdInvitationCreateInput:
      type: object
      properties:
        invitedEmail:
          type: ["string", "null"]
          format: email
        role:
          $ref: "#/components/schemas/HouseholdMemberRole"
        lifecycleState:
          $ref: "#/components/schemas/HouseholdInvitationLifecycleState"
        expiresAt:
          type: ["string", "null"]
          format: date-time
        note:
          type: ["string", "null"]
...
    HouseholdInvitationMutationResponse:
      type: object
      properties:
        invitation:
          $ref: "#/components/schemas/HouseholdInvitation"
        auditEvent:
          $ref: "#/components/schemas/HouseholdAuditEvent"
      required:
        - invitation
        - auditEvent
...
    MeUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
...
    JoinHouseholdInput:
      type: object
      properties:
        inviteCode:
          type: string
          minLength: 1
      required:
        - inviteCode
...
    CareStateEnvelope:
      type: object
      properties:
        householdId:
          type: string
        version:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
          multipleOf: 1
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: ["string", "null"]
        doc:
          type: object
          additionalProperties: true
      required:
        - householdId
        - version
        - updatedAt
        - doc
    CareStateInput:
      type: object
      properties:
        version:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483646
          multipleOf: 1
        doc:
          type: object
          additionalProperties: true
      required:
        - version
        - doc
...